projects and project pitch instr css reorganization
