+++ src/App.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import LoadingAnimation from "../src/components/LoadingAnimation";
 import Navbar from "./components/Navbar";
 import Home from "./components/Home/Home";
-import Projects from "./components/Projects";
+import Projects from "./components/Projects/Projects";
 import FAQ from "./components/Faq";
 import Footer from "./components/Footer";
 import Teams from './components/Team/Team';
