+++ src/components/Home/home.css
@@ -344,14 +344,6 @@
     border-radius: 8px;
 }
 
-/* Project Pitch Section */
-.project-pitch-heading-wrapper {
-    font-size: 2.3em;
-    font-weight: 500;
-    color: white;
-    padding-top: 10px;
-    margin-bottom: 40px;
-}
 
 /* Sponsor Section */
 .sponsor-heading {
