+++ src/components/Projects/ProjectCard.js
@@ -3,6 +3,7 @@ import Card from "react-bootstrap/Card";
 import Button from "react-bootstrap/Button";
 import { CgWebsite } from "react-icons/cg";
 import { BsGithub } from "react-icons/bs";
+import './projects.css';
 
 function ProjectCards(props) {
     return (
