+++ src/components/Schedule/Schedule.js
@@ -90,7 +90,7 @@ const ScheduleCalendar = () => {
   return (
     <Container fluid className="schedule-section">
       <Container>
-        <h1 className="component-heading">Tentative <strong className="purple">Event Schedule</strong></h1>
+        <h1 className="page-heading">Tentative <strong className="purple">Event Schedule</strong></h1>
 
         {isMobile && (
           <div className="date-selector">
